feat: refactor opensource module to follow module pattern and add github stats migration
