+++ server/src/database/prisma/schema/base.prisma.bak
@@ -455,6 +455,7 @@ model opensourceRepo {
   highlights  String[]       @default([])
   trending    Boolean        @default(false)
   lastUpdated DateTime       @default(now())
+  githubStatsUpdatedAt DateTime?
   createdAt            DateTime       @default(now())
   updatedAt            DateTime       @updatedAt
