+++ server/src/module/opensource/opensource.routes.ts
@@ -1,344 +1,58 @@
 import { Router } from "express";
-import { prisma } from "../../database/db.js";
-import {
-  approveRequestOverrideSchema,
-  opensourceListQuerySchema,
-  repoIdSchema,
-  submitRepoRequestSchema,
-} from "./opensource.validation.js";
+import { OpensourceController } from "./opensource.controller.js";
 import { authMiddleware } from "../../middleware/auth.middleware.js";
 import { requireRole } from "../../middleware/role.middleware.js";
-import { sendEmail } from "../../utils/email.utils.js";
-import { repoRequestSubmittedHtml, repoRequestApprovedHtml } from "../../utils/email-templates.js";
-import { parsePagination } from "../../utils/pagination.utils.js";
-import { fetchGithubStats } from "../../lib/github.js";
 
 export const opensourceRouter = Router();
+const controller = new OpensourceController();
 
-function addMonthsUTC(date: Date, months: number): Date {
-  return new Date(Date.UTC(date.getUTCFullYear(), date.getUTCMonth() + months, 1));
-}
+// ─── Public Routes ─────────────────────────────────────────────
 
-function getMonthKeyUTC(date: Date): string {
-  const year = date.getUTCFullYear();
-  const month = String(date.getUTCMonth() + 1).padStart(2, "0");
-  return `${year}-${month}`;
-}
+// List repos with optional filters
+opensourceRouter.get("/", (req, res, next) => controller.listRepos(req, res, next));
 
-function getMonthLabelUTC(date: Date): string {
-  return new Intl.DateTimeFormat("en-US", { month: "short", year: "numeric", timeZone: "UTC" }).format(date);
-}
+// Get all unique languages
+opensourceRouter.get("/languages", (req, res, next) => controller.getLanguages(req, res, next));
 
-// Public: list repos with optional filters
-opensourceRouter.get("/", async (req, res, next) => {
-  try {
-    const parsed = opensourceListQuerySchema.safeParse(req.query);
-    if (!parsed.success) {
-      res.status(400).json({ message: "Invalid query parameters", errors: parsed.error.flatten().fieldErrors });
-      return;
-    }
-    const { page, limit, search, language, difficulty, domain, sortBy, sortOrder } = parsed.data;
+// Get GSoC organizations
+opensourceRouter.get("/gsoc/orgs", (req, res, next) => controller.getGsocOrgs(req, res, next));
 
-    const where: Record<string, unknown> = {};
-    if (language) where["language"] = { equals: language, mode: "insensitive" };
-    if (difficulty) where["difficulty"] = difficulty;
-    if (domain) where["domain"] = domain;
-    if (search) {
-      // Prisma's scalar-list filters can't do case-insensitive substring match
-      // on array elements, so resolve tag matches via a raw ILIKE-on-unnest
-      // subquery and merge the matching ids into the OR clause.
-      const tagMatches = await prisma.$queryRaw<Array<{ id: number }>>`
-        SELECT id FROM "opensourceRepo"
-        WHERE EXISTS (
-          SELECT 1 FROM unnest(tags) AS t WHERE t ILIKE ${`%${search}%`}
-        )
-      `;
-      const tagMatchIds = tagMatches.map((r) => r.id);
-
-      where["OR"] = [
-        { name: { contains: search, mode: "insensitive" } },
-        { owner: { contains: search, mode: "insensitive" } },
-        { description: { contains: search, mode: "insensitive" } },
-        { language: { contains: search, mode: "insensitive" } },
-        ...(tagMatchIds.length > 0 ? [{ id: { in: tagMatchIds } }] : []),
-      ];
-    }
-
-    const skip = (page - 1) * limit;
-    const [repos, total] = await Promise.all([
-      prisma.opensourceRepo.findMany({
-        where,
-        skip,
-        take: limit,
-        orderBy: { [sortBy]: sortOrder },
-      }),
-      prisma.opensourceRepo.count({ where }),
-    ]);
-
-    res.json({
-      repos,
-      pagination: { page, limit, total, totalPages: Math.ceil(total / limit) },
-    });
-  } catch (err) {
-    next(err);
-  }
-});
-
-// ─── Repo Requests (Student-authenticated) ───────────────────────
+// ─── Repo Requests (Student-authenticated) ─────────────────────
 // NOTE: these must be registered BEFORE /:id to avoid route conflicts
 
 // Submit a repo request
-opensourceRouter.post("/requests", authMiddleware, requireRole("STUDENT"), async (req, res, next) => {
-  try {
-    const parsed = submitRepoRequestSchema.safeParse(req.body);
-    if (!parsed.success) {
-      res.status(400).json({ message: "Validation failed", errors: parsed.error.flatten().fieldErrors });
-      return;
-    }
-
-    const existing = await prisma.repoRequest.findFirst({
-      where: { url: parsed.data.url, status: { in: ["PENDING", "APPROVED"] } },
-    });
-    if (existing) {
-      res.status(409).json({ message: "This repository has already been submitted" });
-      return;
-    }
-
-    const request = await prisma.repoRequest.create({
-      data: { ...parsed.data, userId: req.user!.id },
-      include: { user: { select: { name: true, email: true } } },
-    });
-
-    // Send confirmation email
-    try {
-      await sendEmail({
-        to: request.user.email,
-        subject: "Repo Request Received, InternHack",
-        html: repoRequestSubmittedHtml(request.user.name, request.name, request.owner),
-      });
-    } catch { /* email failure is non-blocking */ }
-
-    res.status(201).json({ message: "Repository request submitted successfully", request });
-  } catch (err) {
-    next(err);
-  }
-});
+opensourceRouter.post("/requests", authMiddleware, requireRole("STUDENT"), (req, res, next) => 
+  controller.submitRepoRequest(req, res, next)
+);
 
 // Get my repo requests
-opensourceRouter.get("/requests/mine", authMiddleware, requireRole("STUDENT"), async (req, res, next) => {
-  try {
-    const requests = await prisma.repoRequest.findMany({
-      where: { userId: req.user!.id },
-      orderBy: { createdAt: "desc" },
-    });
-    res.json({ requests });
-  } catch (err) {
-    next(err);
-  }
-});
+opensourceRouter.get("/requests/mine", authMiddleware, requireRole("STUDENT"), (req, res, next) => 
+  controller.getMyRepoRequests(req, res, next)
+);
 
 // Student contribution trend
-opensourceRouter.get("/analytics/trend", authMiddleware, requireRole("STUDENT"), async (req, res, next) => {
-  try {
-    const now = new Date();
-    const currentMonthStart = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), 1));
-    const startMonth = addMonthsUTC(currentMonthStart, -5);
-    const endMonth = addMonthsUTC(currentMonthStart, 1);
+opensourceRouter.get("/analytics/trend", authMiddleware, requireRole("STUDENT"), (req, res, next) => 
+  controller.getStudentContributionTrend(req, res, next)
+);
 
-    const approvedRequests = await prisma.repoRequest.findMany({
-      where: {
-        userId: req.user!.id,
-        status: "APPROVED",
-        updatedAt: { gte: startMonth, lt: endMonth },
-      },
-      select: { updatedAt: true },
-    });
-
-    const countsByMonth = new Map<string, number>();
-    for (const request of approvedRequests) {
-      const monthKey = getMonthKeyUTC(request.updatedAt);
-      countsByMonth.set(monthKey, (countsByMonth.get(monthKey) ?? 0) + 1);
-    }
-
-    const trend = Array.from({ length: 6 }, (_, index) => {
-      const monthStart = addMonthsUTC(startMonth, index);
-      const monthKey = getMonthKeyUTC(monthStart);
-      return {
-        month: monthKey,
-        label: getMonthLabelUTC(monthStart),
-        count: countsByMonth.get(monthKey) ?? 0,
-      };
-    });
-
-    res.json({ trend, total: approvedRequests.length });
-  } catch (err) {
-    next(err);
-  }
-});
-
-// ─── Admin: Manage Repo Requests ─────────────────────────────────
+// ─── Admin: Manage Repo Requests ───────────────────────────────
 
 // List all repo requests
-opensourceRouter.get("/requests/all", authMiddleware, requireRole("ADMIN"), async (req, res, next) => {
-  try {
-    const { page, limit, skip } = parsePagination(req);
-    const status = req.query.status as string | undefined;
-
-    const where: Record<string, unknown> = {};
-    if (status && ["PENDING", "APPROVED", "REJECTED"].includes(status)) {
-      where.status = status;
-    }
-
-    const [requests, total] = await Promise.all([
-      prisma.repoRequest.findMany({
-        where,
-        include: { user: { select: { id: true, name: true, email: true, profilePic: true } } },
-        orderBy: { createdAt: "desc" },
-        skip,
-        take: limit,
-      }),
-      prisma.repoRequest.count({ where }),
-    ]);
-
-    res.json({ requests, pagination: { page, limit, total, totalPages: Math.ceil(total / limit) } });
-  } catch (err) {
-    next(err);
-  }
-});
+opensourceRouter.get("/requests/all", authMiddleware, requireRole("ADMIN"), (req, res, next) => 
+  controller.getAllRepoRequests(req, res, next)
+);
 
 // Approve a repo request
-opensourceRouter.put("/requests/:id/approve", authMiddleware, requireRole("ADMIN"), async (req, res, next) => {
-  try {
-    const id = Number(req.params.id);
-    if (isNaN(id)) { res.status(400).json({ message: "Invalid request ID" }); return; }
-
-    const overridesParsed = approveRequestOverrideSchema.safeParse(req.body);
-    if (!overridesParsed.success) {
-      res.status(400).json({ message: "Validation failed", errors: overridesParsed.error.flatten().fieldErrors });
-      return;
-    }
-    const overrides = overridesParsed.data;
-
-    const request = await prisma.repoRequest.findUnique({
-      where: { id },
-      include: { user: { select: { name: true, email: true } } },
-    });
-    if (!request) { res.status(404).json({ message: "Request not found" }); return; }
-    if (request.status !== "PENDING") { res.status(400).json({ message: "Request is not pending" }); return; }
-
-    // Create the actual repo from the request
-    const repo = await prisma.opensourceRepo.create({
-      data: {
-        name: overrides.name ?? request.name,
-        owner: request.owner,
-        description: overrides.description ?? request.description,
-        language: request.language,
-        url: request.url,
-        domain: overrides.domain ?? request.domain,
-        difficulty: overrides.difficulty ?? request.difficulty,
-        techStack: request.techStack,
-        tags: overrides.tags ?? request.tags,
-      },
-    });
-
-    // Update request status
-    await prisma.repoRequest.update({
-      where: { id },
-      data: { status: "APPROVED", adminNote: overrides.adminNote ?? null },
-    });
-
-    // Send approval email
-    try {
-      await sendEmail({
-        to: request.user.email,
-        subject: "Your Repo Has Been Approved, InternHack",
-        html: repoRequestApprovedHtml(request.user.name, overrides.name ?? request.name, request.owner),
-      });
-    } catch { /* email failure is non-blocking */ }
-
-    // Fetch stats on approval
-    fetchGithubStats(repo.url)
-      .then(async (stats) => {
-        if (!stats) return;
-        await prisma.opensourceRepo.update({
-          where: { id: repo.id },
-          data: {
-            stars: stats.stars,
-            forks: stats.forks,
-            openIssues: stats.openIssues,
-            githubStatsUpdatedAt: new Date(),
-          } as any,
-        });
-      })
-      .catch((err) => console.error("[github] approval stats fetch failed:", err));
-
-    res.json({ message: "Request approved and repository added", repo });
-  } catch (err) {
-    next(err);
-  }
-});
+opensourceRouter.put("/requests/:id/approve", authMiddleware, requireRole("ADMIN"), (req, res, next) => 
+  controller.approveRepoRequest(req, res, next)
+);
 
 // Reject a repo request
-opensourceRouter.put("/requests/:id/reject", authMiddleware, requireRole("ADMIN"), async (req, res, next) => {
-  try {
-    const id = Number(req.params.id);
-    if (isNaN(id)) { res.status(400).json({ message: "Invalid request ID" }); return; }
+opensourceRouter.put("/requests/:id/reject", authMiddleware, requireRole("ADMIN"), (req, res, next) => 
+  controller.rejectRepoRequest(req, res, next)
+);
 
-    const request = await prisma.repoRequest.findUnique({ where: { id } });
-    if (!request) { res.status(404).json({ message: "Request not found" }); return; }
-    if (request.status !== "PENDING") { res.status(400).json({ message: "Request is not pending" }); return; }
-
-    await prisma.repoRequest.update({
-      where: { id },
-      data: { status: "REJECTED", adminNote: req.body.adminNote || null },
-    });
-
-    res.json({ message: "Request rejected" });
-  } catch (err) {
-    next(err);
-  }
-});
+// ─── Public: Single Repo ───────────────────────────────────────
 
 // Public: get single repo (must be AFTER /requests/* routes)
-opensourceRouter.get("/:id", async (req, res, next) => {
-  try {
-    const parsed = repoIdSchema.safeParse(req.params);
-    if (!parsed.success) { res.status(400).json({ message: "Invalid repo ID" }); return; }
-    const { id } = parsed.data;
-
-    const repo = (await prisma.opensourceRepo.findUnique({ where: { id } })) as any;
-    if (!repo) { res.status(404).json({ message: "Repository not found" }); return; }
-
-    // Start: Background stats update (non-blocking)
-    const SIX_HOURS = 6 * 60 * 60 * 1000;
-    const isStale =
-      !repo.githubStatsUpdatedAt ||
-      Date.now() - new Date(repo.githubStatsUpdatedAt).getTime() > SIX_HOURS;
-
-    if (isStale && repo.url?.includes("github.com")) {
-      fetchGithubStats(repo.url)
-        .then(async (stats) => {
-          if (!stats) return;
-          await prisma.opensourceRepo.update({
-            where: { id: repo.id },
-            data: {
-              stars: stats.stars,
-              forks: stats.forks,
-              openIssues: stats.openIssues,
-              githubStatsUpdatedAt: new Date(),
-              ...(stats.language && { language: stats.language }),
-            } as any,
-          });
-          console.info(`[github] updated stats for ${repo.name}`);
-        })
-        .catch((err) =>
-          console.error(`[github] background update failed for ${repo.id}:`, err)
-        );
-    }
-    // End: Background stats update
-
-    res.json({ repo });
-  } catch (err) {
-    next(err);
-  }
-});
+opensourceRouter.get("/:id", (req, res, next) => controller.getRepoById(req, res, next));
